WordClound/250426_ice_tourism_wordcloud/requirements.txt
jieba
wordcloud
matplotlib
numpy
pillow